add config parsing to config module
